// acad-front/src/app/category/models/antedecedente-medico-model.ts
export class AntecedenteMedicoModel {
  idAntecedenteMedico?: number;
  edad?: string;
  altura?: string;
  peso?: string;
  grupoSanguineo?: string;
  enfermedadesPadecidas?: string;
  antecedenteDeHospitalizacionoDeCirugias?: string;
  seEncuentraMedicado?: string;
  presentaDiscapacidad?: string;
  alergias?: string;
  estadoPsicologico?: string;
}